requirements.txt
google-genai>=0.8.0
Pillow>=10.0.0
python-dotenv>=1.0.0
Flask>=2.3.0
Flask-CORS>=4.0.0
Werkzeug>=2.3.0
numpy>=1.21.0
requests>=2.28.0
boto3>=1.28.0
